build_simulations is again the single backwards-compatible entry point for all cases including mts. Tried to simplify the preprocessing script
