+++ scripts/build_and_run_mid_term_scheduling.py
@@ -19,22 +19,22 @@ import dispaset as ds
 config = ds.load_config('../ConfigFiles/Config_Mid_Term_Scheduling.xlsx')
 
 ## Build simulation data with new profiles
-SimData = ds.build_full_simulation(config,mts_plot=True)
+SimData = ds.build_simulation(config, mts_plot=True)
 
 #Solve using GAMS:
-r = ds.solve_GAMS(config['SimulationDirectory'], config['GAMS_folder'])
+#r = ds.solve_GAMS(config['SimulationDirectory'], config['GAMS_folder'])
 
 # Load the simulation results:
-inputs,results = ds.get_sim_results(config['SimulationDirectory'],cache=False)
+#inputs,results = ds.get_sim_results(config['SimulationDirectory'],cache=False)
 
 # Generate country-specific plots
-ds.plot_zone(inputs,results,z='DE')
+#ds.plot_zone(inputs,results,z='DE')
 
 # Bar plot with the installed capacities in all countries:
-cap = ds.plot_zone_capacities(inputs)
+#cap = ds.plot_zone_capacities(inputs)
 
 # Bar plot with the energy balances in all countries:
-ds.plot_energy_zone_fuel(inputs,results,ds.get_indicators_powerplant(inputs,results))
+#ds.plot_energy_zone_fuel(inputs,results,ds.get_indicators_powerplant(inputs,results))
 
 # Analyse the results for each country and provide quantitative indicators:
-r = ds.get_result_analysis(inputs,results)
+#r = ds.get_result_analysis(inputs,results)
